Move and refactor handleChanges API
